invoice changes to support multiple items on the document
Invoice edit view moved from modal to separated view. In future a modal
will be used to create or modify items on the invoice.

Along the way I found multiple issues related to calculation of total
values from the list of items caused by using for loop on the array
improperly. Added tests and confirmed it working.

Only changes are necessary is UI view with table of invoices and remove
hardcodes that adds first element.