pdf generation font picker
Commit introduces two new dropdowns related to fonts in config section
- font family
font picker config
That commit introduces a new feature, which allows to pick any font you'd like
to to use in your invoices. It was kind of unfortunate to follow the hardcoded
JetBrainsMono that I personally like. Bunch of places have been redesigned
along the way mostly related to PDF area.

For now fc-list has been used to list all the system fonts but it is not ideal.
I have to figure a way to make it more generic and system agnostic.