Print invoice security adjustments
Since failure modal got introduced this change adds further details to
why generating PDF failed. Most of the reasons is the template has not
enough field size to fill up the content and that should be addressed on
either template or invoice data level.

Moreover most of the value replacements got refactored towards pointer
which I find more elegant and easier to work with in loop and way more
convenient in case of returning errors.